voice bot/backend/requirements.txt
fastapi
uvicorn
websockets
google-generativeai
dotenv
gTTS
